added comments for components, moved transforms to own file
